Code Optimzation: Article Add Update
