+++ lib/admin/frontend/view/dashboard.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 
 import 'package:google_fonts/google_fonts.dart';
 import 'package:rafay_portfolio/admin/backend/auth/authService.dart';
-import 'package:rafay_portfolio/admin/frontend/pages/article/BlogsWork.dart';
+import 'package:rafay_portfolio/admin/frontend/pages/article/articles.dart';
 import 'package:rafay_portfolio/admin/frontend/view/MessageDashboard.dart';
 import 'package:rafay_portfolio/admin/frontend/widgets/admin_drawer.dart';
