+++ pubspec.yaml
@@ -33,7 +33,6 @@ dependencies:
   image_picker_web: ^3.1.1
   flutter_html: ^3.0.0-beta.2
   flutter_animation_progress_bar: ^2.3.1
-  markdown_editor_plus: ^0.2.13
   firebase_crashlytics: ^3.4.19
   timeline_tile: ^2.0.0
   showcaseview: ^2.0.3
@@ -42,6 +41,7 @@ dependencies:
   scroll_to_index: ^3.0.1
   cached_network_image: ^3.3.1
   rxdart: ^0.27.7
+  flutter_markdown: ^0.6.22
 dev_dependencies:
   flutter_test:
     sdk: flutter
